Set up process.env
