+++ package.json
@@ -99,6 +99,8 @@
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.1",
+    "dotenv": "^8.2.0",
+    "dotenv-expand": "^5.1.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.0.0-beta.1",
     "mini-css-extract-plugin": "^1.3.3",
